🛠️Refactor: PR 피드백 반영
