+++ src/main/java/com/bombombom/devs/global/security/JwtAuthenticationFilter.java
@@ -54,7 +54,6 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
                 userDetails.getPassword(),
                 userDetails.getAuthorities()
             );
-            authentication.setDetails(userDetails);
             SecurityContextHolder.getContext().setAuthentication(authentication);
         }
