+++ src/main/java/com/bombombom/devs/study/controller/StudyController.java
@@ -7,8 +7,8 @@ import com.bombombom.devs.study.controller.dto.request.RegisterBookStudyRequest;
 import com.bombombom.devs.study.controller.dto.response.StudyPageResponse;
 import com.bombombom.devs.study.service.StudyService;
 import com.bombombom.devs.global.security.AppUserDetails;
+import jakarta.validation.Valid;
 import java.net.URI;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
@@ -32,7 +32,7 @@ public class StudyController {
 
     @PostMapping("/algo")
     public ResponseEntity<Void> registerAlgorithmStudy(
-        @RequestBody RegisterAlgorithmStudyRequest registerAlgorithmStudyRequest) {
+        @Valid @RequestBody RegisterAlgorithmStudyRequest registerAlgorithmStudyRequest) {
         log.info("{}", registerAlgorithmStudyRequest);
         Long id = studyService.createAlgorithmStudy(registerAlgorithmStudyRequest.toServiceDto());
         return ResponseEntity.created(URI.create(RESOURCE_PATH + "/" + id)).build();
@@ -40,7 +40,7 @@ public class StudyController {
 
     @PostMapping("/book")
     public ResponseEntity<Void> registerBookStudy(
-        @RequestBody RegisterBookStudyRequest registerBookStudyRequest) {
+        @Valid @RequestBody RegisterBookStudyRequest registerBookStudyRequest) {
         log.info("{}", registerBookStudyRequest);
         Long id = studyService.createBookStudy(registerBookStudyRequest.toServiceDto());
         return ResponseEntity.created(URI.create(RESOURCE_PATH + id)).build();
@@ -54,9 +54,9 @@ public class StudyController {
     }
 
     @PostMapping("/join")
-    public ResponseEntity<Void> joinAlgorithmStudy(@LoginUser AppUserDetails userDetails,
-        @RequestBody JoinStudyRequest joinStudyRequest) {
-        studyService.joinAlgorithmStudy(userDetails.getId(), joinStudyRequest.toServiceDto());
+    public ResponseEntity<Void> joinStudy(@LoginUser AppUserDetails userDetails,
+        @Valid @RequestBody JoinStudyRequest joinStudyRequest) {
+        studyService.joinStudy(userDetails.getId(), joinStudyRequest.toServiceDto());
         return ResponseEntity.ok().build();
     }
